Basic setup
Set up basic server and database connection with Express and Mongoose
