+++ nodemon.json
@@ -1,5 +1,5 @@
 {
     "watch": ["src"], 
     "ext":".ts,.js",
-    "exec": "ts-node ./src/index.ts"
+    "exec": "ts-node ./src/server.ts"
 }
