+++ src/db.ts
@@ -1,0 +1,11 @@
+import mongoose from 'mongoose';
+
+export const connectDB = async () => {
+  try {
+    await mongoose.connect(process.env.MONGO_URI!);
+    console.log('Connected to MongoDB');
+  } catch (err) {
+    console.error('Failed to connect to MongoDB:', err);
+    process.exit(1);
+  }
+};
